Feature/134 validation (#135)
* Removed FLuentValidation

* Removed unused usings from ApplicationModule

* Removed query/command references from controllers

* Removed ReSharper overrides

* Added ModelValidator

* Added basic stub validator and some basic unit tests

* Added validation for response status code

* Moved config models to application project

* Moved ConfigurationParser to infrastructure project and removed Configuration projects

* Moved ConfigurationParser to the correct namespace

* Added config key for extraDuration millis and fixed some other stuff

* Did some extra refactoring and implemented extraDuration check + unit tests

* Added enum for working with line endings

* Added failing unit tests for line ending type validation

* Implemented validation for lineEndings

* Added enum for response image type

* Added validations for image types: colors and JPEG quality

* Added new failing scenario to StubModelValidatorFacts

* Fixed issue with status code validation if status code is 0

* Updated StatusCode validation rule and added validation to YamlFileStubSource

* Param name change

* Added validation to AddStubCommandHandler and added unit tests

* Added validation to UpdateStubCommandHandler and added unit tests

* Fixed a test and added message to ValidationException

* Added integration test for validation exception

* Added IT to Postman collection

* Small fix in Postman test runner wrapper script

* Added HTTP 400 error handling to UI

* Simplified StubForm

+++ src/HttPlaceholder.Application/Stubs/Commands/AddStub/AddStubCommandHandler.cs
@@ -1,23 +1,32 @@
-﻿using System.Threading;
+﻿using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
+using HttPlaceholder.Application.Exceptions;
 using HttPlaceholder.Application.StubExecution;
 using HttPlaceholder.Domain;
 using MediatR;
 
 namespace HttPlaceholder.Application.Stubs.Commands.AddStub
 {
-    // ReSharper disable once UnusedType.Global
     public class AddStubCommandHandler : IRequestHandler<AddStubCommand, FullStubModel>
     {
         private readonly IStubContext _stubContext;
+        private readonly IStubModelValidator _stubModelValidator;
 
-        public AddStubCommandHandler(IStubContext stubContext)
+        public AddStubCommandHandler(IStubContext stubContext, IStubModelValidator stubModelValidator)
         {
             _stubContext = stubContext;
+            _stubModelValidator = stubModelValidator;
         }
 
         public async Task<FullStubModel> Handle(AddStubCommand request, CancellationToken cancellationToken)
         {
+            var validationResults = _stubModelValidator.ValidateStubModel(request.Stub);
+            if (validationResults.Any())
+            {
+                throw new ValidationException(validationResults);
+            }
+
             // Delete stub with same ID.
             await _stubContext.DeleteStubAsync(request.Stub.Id);
             return await _stubContext.AddStubAsync(request.Stub);
